abstracted graph
